+++ server/src/main/java/com/lankacapital/server/dtos/LoanCreateDto.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 public class LoanCreateDto {
     private String fileNumber;
     private BigDecimal loanAmount;
-    private Integer interestRate;
+    private Double interestRate;
     private BigDecimal documentCharge;
     private Integer numberOfInstallments;
